requirements.txt
django-crispy-forms==1.7.2
django==2.1.5
pytest==4.1.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.1
